# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SwiftEntry API",
        "description": "Backend REST de SwiftEntry \u2014 plataforma de venta de boletos para eventos.\n\n**Autenticaci\u00f3n:** todas las rutas protegidas requieren un JWT en el header `Authorization: Bearer <token>`. Obt\u00e9n el token en `POST /swift_entry/auth/login` y p\u00e9galo en el bot\u00f3n **Authorize** de arriba.\n\n**Flujo principal:** Evento \u2192 Localidad \u2192 Asiento \u2192 Reserva \u2192 Pago \u2192 Boleto\n",
        "contact": {
            "name": "SwiftEntry Team",
            "email": "gerardo.port1107.psql@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/swift_entry/events/{id}": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Obtener evento por ID",
                "operationId": "getEventById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Actualizar evento",
                "operationId": "updateEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventUpdateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Eliminar evento",
                "operationId": "deleteEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/swift_entry/events": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Listar todos los eventos",
                "description": "Ruta p\u00fablica",
                "operationId": "getAllEvents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Crear evento",
                "description": "Ruta p\u00fablica. Crea el evento junto con sus localidades",
                "operationId": "createEvent",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EventRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/swift_entry/addresses": {
            "get": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Listar todas las direcciones",
                "operationId": "getAddress",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddressRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Crear direcci\u00f3n",
                "operationId": "createAddress",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddressRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/swift_entry/seats/event/{eventId}": {
            "get": {
                "tags": [
                    "Asientos"
                ],
                "summary": "Mapa de asientos de un evento",
                "description": "Devuelve todos los locality_seats del evento con su estado actual (AVAILABLE/RESERVED/SOLD)",
                "operationId": "getSeatMapByEventId",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/swift_entry/localities/event/{eventId}": {
            "get": {
                "tags": [
                    "Localidades"
                ],
                "summary": "Localidades de un evento",
                "operationId": "getLocalitiesByEventId",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/swift_entry/events/organizer/{id}": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Eventos por organizador",
                "operationId": "getAllEventsByOrganizerId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "LocalityUpdateDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "minimum": 0.0
                    },
                    "earlyBirdDiscountPercentage": {
                        "type": "number",
                        "maximum": 100.0,
                        "minimum": 0.0
                    },
                    "earlyBirdDeadline": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "capacity": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "availableSlots": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    }
                }
            },
            "GeneralResponse": {
                "type": "object",
                "properties": {
                    "uri": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "time": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "data": {}
                }
            },
            "EventUpdateDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 150,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 0
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "CONCERT",
                            "SPORTS",
                            "THEATER",
                            "COMEDY",
                            "CONFERENCE",
                            "FESTIVAL",
                            "MOVIE",
                            "CULTURAL"
                        ]
                    },
                    "organizerId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "venueName": {
                        "type": "string",
                        "maxLength": 150,
                        "minLength": 0
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "PUBLISHED",
                            "CANCELLED",
                            "FINISHED"
                        ]
                    },
                    "imageUrl": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 0
                    },
                    "localities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LocalityUpdateDTO"
                        }
                    }
                }
            },
            "AddressRequestDTO": {
                "type": "object",
                "properties": {
                    "streetAddress": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "neighborhood": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "municipality": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "department": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "country": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "referencePoint": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 0
                    }
                },
                "required": [
                    "country",
                    "department",
                    "municipality",
                    "streetAddress"
                ]
            },
...
            "LocalityRequestDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "minimum": 0.0
                    },
                    "earlyBirdDiscountPercentage": {
                        "type": "number",
                        "maximum": 100.0,
                        "minimum": 0.0
                    },
                    "earlyBirdDeadline": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "name",
                    "price"
                ]
            },
            "EventRequestDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 150,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "minLength": 0
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "CONCERT",
                            "SPORTS",
                            "THEATER",
                            "COMEDY",
                            "CONFERENCE",
                            "FESTIVAL",
                            "MOVIE",
                            "CULTURAL"
                        ]
                    },
                    "organizerId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "venueName": {
                        "type": "string",
                        "maxLength": 150,
                        "minLength": 0
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "PUBLISHED",
                            "CANCELLED",
                            "FINISHED"
                        ]
                    },
                    "imageUrl": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 0
                    },
                    "localities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LocalityRequestDTO"
                        },
                        "maxItems": 2147483647,
                        "minItems": 1
                    }
                },
                "required": [
                    "category",
                    "endDate",
                    "localities",
                    "name",
                    "organizerId",
                    "startDate",
                    "status"
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}